# docs/api-contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nivesh API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/funds/": {
      "post": {
        "tags": [
          "funds"
        ],
        "summary": "Create Fund",
        "description": "Create a new fund entry (Requires Auth).",
        "operationId": "create_fund_api_v1_funds__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FundMasterCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundMasterRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "funds"
        ],
        "summary": "List Funds",
        "description": "List all mutual funds with pagination and filtering.",
        "operationId": "list_funds_api_v1_funds__get",
        "parameters": [
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by active status",
              "title": "Is Active"
            },
            "description": "Filter by active status"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by scheme category",
              "title": "Category"
            },
            "description": "Filter by scheme category"
          },
          {
            "name": "subcategory",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by scheme subcategory",
              "title": "Subcategory"
            },
            "description": "Filter by scheme subcategory"
          },
          {
            "name": "amc",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by AMC name",
              "title": "Amc"
            },
            "description": "Filter by AMC name"
          },
          {
            "name": "plan_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by plan type (Direct/Regular)",
              "title": "Plan Type"
            },
            "description": "Filter by plan type (Direct/Regular)"
          },
          {
            "name": "benchmark_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by benchmark index code",
              "title": "Benchmark Code"
            },
            "description": "Filter by benchmark index code"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search by fund name or code",
              "title": "Q"
            },
            "description": "Search by fund name or code"
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sort field (scheme_name, -scheme_name)",
              "default": "scheme_name",
              "title": "Order By"
            },
            "description": "Sort field (scheme_name, -scheme_name)"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundMasterListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/funds/categories/{category}/subcategories": {
      "get": {
        "tags": [
          "funds"
        ],
        "summary": "List Subcategories",
        "description": "Return distinct subcategories for a given category.",
        "operationId": "list_subcategories_api_v1_funds_categories__category__subcategories_get",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response List Subcategories Api V1 Funds Categories  Category  Subcategories Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/funds/compare": {
      "get": {
        "tags": [
          "funds"
        ],
        "summary": "Compare Funds",
        "description": "Compare multiple funds (max 4). Enforces same category matching.",
        "operationId": "compare_funds_api_v1_funds_compare_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "codes",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Comma-separated scheme codes for comparison",
              "title": "Codes"
            },
            "description": "Comma-separated scheme codes for comparison"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/funds/{scheme_code}": {
      "get": {
        "tags": [
          "funds"
        ],
        "summary": "Read Fund",
        "description": "Get details for a specific fund by its code.",
        "operationId": "read_fund_api_v1_funds__scheme_code__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scheme_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scheme Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundMasterRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "funds"
        ],
        "summary": "Update Fund",
        "description": "Update fund details (Requires Auth).",
        "operationId": "update_fund_api_v1_funds__scheme_code__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scheme_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scheme Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FundMasterUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundMasterRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "funds"
        ],
        "summary": "Delete Fund",
        "description": "Delete (deactivate) a fund (Requires Auth).",
        "operationId": "delete_fund_api_v1_funds__scheme_code__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scheme_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scheme Code"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/funds/{scheme_code}/similar": {
      "get": {
        "tags": [
          "funds"
        ],
        "summary": "Get Similar Funds",
        "description": "Get funds with the same category and subcategory.",
        "operationId": "get_similar_funds_api_v1_funds__scheme_code__similar_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scheme_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scheme Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FundMasterRead"
                  },
                  "title": "Response Get Similar Funds Api V1 Funds  Scheme Code  Similar Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/stocks/{symbol}": {
      "get": {
        "tags": [
          "Stocks"
        ],
        "summary": "Get Stock",
        "operationId": "get_stock_api_v1_stocks__symbol__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FundMasterListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/FundMasterRead"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "skip",
          "limit",
          "items"
        ],
        "title": "FundMasterListResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}